+++ database/migrations/2019_08_10_173706_create_healths_table.php
@@ -15,7 +15,7 @@ class CreateHealthsTable extends Migration
     {
         Schema::create('healths', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('characterId')->unsigned();
+            $table->bigInteger('characterId')->unsigned();
             $table->foreign('characterId')->references('id')->on('characters');
             $table->tinyInteger('position');
             $table->string('value', 10);
